+++ api/components/auth/network.js
@@ -8,10 +8,17 @@ const router = express.Router()
 require('../../../auth/strategies/basic')
 
 router.post('/login', passport.authenticate('basic', { session: false }), token)
+router.post('/register', register)
 
 function token (req, res) {
   Controller.token(req.user)
     .then(token => response.success(req, res, token, 200))
+    .catch(() => response.error(req, res, 'Invalid information', 400))
+}
+
+function register (req, res) {
+  Controller.register(req.body)
+    .then(status => response.success(req, res, status, 200))
     .catch(() => response.error(req, res, 'Invalid information', 400))
 }
