+++ api/components/user/network.js
@@ -1,5 +1,4 @@
 const express = require('express')
-const passport = require('passport')
 const router = express.Router()
 
 const response = require('../../../network/response')
@@ -8,18 +7,11 @@ const Controller = require('./index')
 // JWT Strategy
 require('../../../auth/strategies/jwt')
 
-router.get('/', passport.authenticate('jwt', { session: false }), list)
-router.post('/', register)
+router.get('/', list)
 
 function list (req, res, next) {
   Controller.list()
     .then(list => response.success(req, res, list, 200))
-    .catch(next)
-}
-
-function register (req, res, next) {
-  Controller.register(req.body)
-    .then(user => response.success(req, res, user, 201))
     .catch(next)
 }
