+++ package.json
@@ -15,6 +15,7 @@
     "@hapi/boom": "^9.1.0",
     "bcrypt": "^4.0.1",
     "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
